:zap: logic split
