+++ packages/nodes-base/nodes/ItemLists/V3/actions/itemList/sort.operation.ts
@@ -1,6 +1,17 @@
 import type { IExecuteFunctions } from 'n8n-core';
 import type { IDataObject, INodeExecutionData, INodeProperties } from 'n8n-workflow';
-import { updateDisplayOptions, wrapData } from '@utils/utilities';
+import { NodeOperationError } from 'n8n-workflow';
+import { updateDisplayOptions } from '@utils/utilities';
+
+import type { NodeVMOptions } from 'vm2';
+import { NodeVM } from 'vm2';
+
+import get from 'lodash/get';
+
+import isEqual from 'lodash/isEqual';
+import lt from 'lodash/lt';
+
+import { shuffleArray } from '../../helpers/utils';
 
 const properties: INodeProperties[] = [
 	{
@@ -145,24 +156,156 @@ export async function execute(
 	this: IExecuteFunctions,
 	items: INodeExecutionData[],
 ): Promise<INodeExecutionData[]> {
-	const returnData: INodeExecutionData[] = [];
+	let returnData = [...items];
+	const type = this.getNodeParameter('type', 0) as string;
+	const disableDotNotation = this.getNodeParameter(
+		'options.disableDotNotation',
+		0,
+		false,
+	) as boolean;
 
-	for (let i = 0; i < items.length; i++) {
-		try {
-			const data: IDataObject[] = [];
-			const executionData = this.helpers.constructExecutionMetaData(wrapData(data), {
-				itemData: { item: i },
-			});
-
-			returnData.push(...executionData);
-		} catch (error) {
-			if (this.continueOnFail()) {
-				returnData.push({ json: { error: error.message } });
-				continue;
-			}
-			throw error;
-		}
+	if (type === 'random') {
+		shuffleArray(returnData);
+		return returnData;
 	}
 
+	if (type === 'simple') {
+		const sortFieldsUi = this.getNodeParameter('sortFieldsUi', 0) as IDataObject;
+		const sortFields = sortFieldsUi.sortField as Array<{
+			fieldName: string;
+			order: 'ascending' | 'descending';
+		}>;
+
+		if (!sortFields?.length) {
+			throw new NodeOperationError(
+				this.getNode(),
+				'No sorting specified. Please add a field to sort by',
+			);
+		}
+
+		for (const { fieldName } of sortFields) {
+			let found = false;
+			for (const item of items) {
+				if (!disableDotNotation) {
+					if (get(item.json, fieldName) !== undefined) {
+						found = true;
+					}
+				} else if (item.json.hasOwnProperty(fieldName)) {
+					found = true;
+				}
+			}
+			if (!found && disableDotNotation && fieldName.includes('.')) {
+				throw new NodeOperationError(
+					this.getNode(),
+					`Couldn't find the field '${fieldName}' in the input data`,
+					{
+						description:
+							"If you're trying to use a nested field, make sure you turn off 'disable dot notation' in the node options",
+					},
+				);
+			} else if (!found) {
+				throw new NodeOperationError(
+					this.getNode(),
+					`Couldn't find the field '${fieldName}' in the input data`,
+				);
+			}
+		}
+
+		const sortFieldsWithDirection = sortFields.map((field) => ({
+			name: field.fieldName,
+			dir: field.order === 'ascending' ? 1 : -1,
+		}));
+
+		returnData.sort((a, b) => {
+			let result = 0;
+			for (const field of sortFieldsWithDirection) {
+				let equal;
+				if (!disableDotNotation) {
+					const _a =
+						typeof get(a.json, field.name) === 'string'
+							? (get(a.json, field.name) as string).toLowerCase()
+							: get(a.json, field.name);
+					const _b =
+						typeof get(b.json, field.name) === 'string'
+							? (get(b.json, field.name) as string).toLowerCase()
+							: get(b.json, field.name);
+					equal = isEqual(_a, _b);
+				} else {
+					const _a =
+						typeof a.json[field.name] === 'string'
+							? (a.json[field.name] as string).toLowerCase()
+							: a.json[field.name];
+					const _b =
+						typeof b.json[field.name] === 'string'
+							? (b.json[field.name] as string).toLowerCase()
+							: b.json[field.name];
+					equal = isEqual(_a, _b);
+				}
+
+				if (!equal) {
+					let lessThan;
+					if (!disableDotNotation) {
+						const _a =
+							typeof get(a.json, field.name) === 'string'
+								? (get(a.json, field.name) as string).toLowerCase()
+								: get(a.json, field.name);
+						const _b =
+							typeof get(b.json, field.name) === 'string'
+								? (get(b.json, field.name) as string).toLowerCase()
+								: get(b.json, field.name);
+						lessThan = lt(_a, _b);
+					} else {
+						const _a =
+							typeof a.json[field.name] === 'string'
+								? (a.json[field.name] as string).toLowerCase()
+								: a.json[field.name];
+						const _b =
+							typeof b.json[field.name] === 'string'
+								? (b.json[field.name] as string).toLowerCase()
+								: b.json[field.name];
+						lessThan = lt(_a, _b);
+					}
+					if (lessThan) {
+						result = -1 * field.dir;
+					} else {
+						result = 1 * field.dir;
+					}
+					break;
+				}
+			}
+			return result;
+		});
+	} else {
+		const code = this.getNodeParameter('code', 0) as string;
+		const regexCheck = /\breturn\b/g.exec(code);
+
+		if (regexCheck?.length) {
+			const sandbox = {
+				newItems: returnData,
+			};
+			const mode = this.getMode();
+			const options = {
+				console: mode === 'manual' ? 'redirect' : 'inherit',
+				sandbox,
+			};
+			const vm = new NodeVM(options as unknown as NodeVMOptions);
+
+			returnData = await vm.run(
+				`
+			module.exports = async function() {
+				newItems.sort( (a,b) => {
+					${code}
+				})
+				return newItems;
+			}()`,
+				__dirname,
+			);
+		} else {
+			throw new NodeOperationError(
+				this.getNode(),
+				"Sort code doesn't return. Please add a 'return' statement to your code",
+			);
+		}
+	}
 	return returnData;
 }
