+++ packages/nodes-base/nodes/ItemLists/V3/actions/itemList/splitOutItems.operation.ts
@@ -1,6 +1,11 @@
 import type { IExecuteFunctions } from 'n8n-core';
 import type { IDataObject, INodeExecutionData, INodeProperties } from 'n8n-workflow';
-import { updateDisplayOptions, wrapData } from '@utils/utilities';
+import { deepCopy, NodeOperationError } from 'n8n-workflow';
+
+import { updateDisplayOptions } from '@utils/utilities';
+
+import get from 'lodash/get';
+import unset from 'lodash/unset';
 
 const properties: INodeProperties[] = [
 	{
@@ -110,19 +115,126 @@ export async function execute(
 	const returnData: INodeExecutionData[] = [];
 
 	for (let i = 0; i < items.length; i++) {
-		try {
-			const data: IDataObject[] = [];
-			const executionData = this.helpers.constructExecutionMetaData(wrapData(data), {
-				itemData: { item: i },
-			});
+		const fieldsToSplitOut = (this.getNodeParameter('fieldToSplitOut', i) as string)
+			.split(',')
+			.map((field) => field.trim());
+		const disableDotNotation = this.getNodeParameter(
+			'options.disableDotNotation',
+			0,
+			false,
+		) as boolean;
 
-			returnData.push(...executionData);
-		} catch (error) {
-			if (this.continueOnFail()) {
-				returnData.push({ json: { error: error.message } });
-				continue;
+		const destinationFields = (
+			this.getNodeParameter('options.destinationFieldName', i, '') as string
+		)
+			.split(',')
+			.filter((field) => field.trim() !== '')
+			.map((field) => field.trim());
+
+		if (destinationFields.length && destinationFields.length !== fieldsToSplitOut.length) {
+			throw new NodeOperationError(
+				this.getNode(),
+				'If multiple fields to split out are given, the same number of destination fields must be given',
+			);
+		}
+
+		const include = this.getNodeParameter('include', i) as
+			| 'selectedOtherFields'
+			| 'allOtherFields'
+			| 'noOtherFields';
+
+		const multiSplit = fieldsToSplitOut.length > 1;
+
+		const item = { ...items[i].json };
+		const splited: IDataObject[] = [];
+		for (const [entryIndex, fieldToSplitOut] of fieldsToSplitOut.entries()) {
+			const destinationFieldName = destinationFields[entryIndex] || '';
+
+			let arrayToSplit;
+			if (!disableDotNotation) {
+				arrayToSplit = get(item, fieldToSplitOut);
+			} else {
+				arrayToSplit = item[fieldToSplitOut];
 			}
-			throw error;
+
+			if (arrayToSplit === undefined) {
+				arrayToSplit = [];
+			}
+
+			if (typeof arrayToSplit !== 'object' || arrayToSplit === null) {
+				arrayToSplit = [arrayToSplit];
+			}
+
+			if (!Array.isArray(arrayToSplit)) {
+				arrayToSplit = Object.values(arrayToSplit);
+			}
+
+			for (const [elementIndex, element] of arrayToSplit.entries()) {
+				if (splited[elementIndex] === undefined) {
+					splited[elementIndex] = {};
+				}
+
+				const fieldName = destinationFieldName || fieldToSplitOut;
+
+				if (typeof element === 'object' && element !== null && include === 'noOtherFields') {
+					if (destinationFieldName === '' && !multiSplit) {
+						splited[elementIndex] = { ...splited[elementIndex], ...element };
+					} else {
+						splited[elementIndex][fieldName] = element;
+					}
+				} else {
+					splited[elementIndex][fieldName] = element;
+				}
+			}
+		}
+
+		for (const splitEntry of splited) {
+			let newItem: IDataObject = {};
+
+			if (include === 'noOtherFields') {
+				newItem = splitEntry;
+			}
+
+			if (include === 'allOtherFields') {
+				const itemCopy = deepCopy(item);
+				for (const fieldToSplitOut of fieldsToSplitOut) {
+					if (!disableDotNotation) {
+						unset(itemCopy, fieldToSplitOut);
+					} else {
+						delete itemCopy[fieldToSplitOut];
+					}
+				}
+				newItem = { ...itemCopy, ...splitEntry };
+			}
+
+			if (include === 'selectedOtherFields') {
+				const fieldsToInclude = (
+					this.getNodeParameter('fieldsToInclude.fields', i, []) as [{ fieldName: string }]
+				).map((field) => field.fieldName);
+
+				if (!fieldsToInclude.length) {
+					throw new NodeOperationError(this.getNode(), 'No fields specified', {
+						description: 'Please add a field to include',
+					});
+				}
+
+				for (const field of fieldsToInclude) {
+					if (!disableDotNotation) {
+						splitEntry[field] = get(item, field);
+					} else {
+						splitEntry[field] = item[field];
+					}
+				}
+
+				newItem = splitEntry;
+			}
+
+			returnData.push({
+				json: newItem,
+				pairedItem: {
+					item: i,
+				},
+			});
 		}
 	}
