+++ packages/nodes-base/nodes/ItemLists/V3/actions/itemList/summarize.operation.ts
@@ -539,7 +539,7 @@ function aggregationToArray(
 export async function execute(
 	this: IExecuteFunctions,
 	items: INodeExecutionData[],
-): Promise<INodeExecutionData[][]> {
+): Promise<INodeExecutionData[]> {
 	const newItems = items.map(({ json }, i) => ({ ...json, _itemIndex: i }));
 
 	const options = this.getNodeParameter('options', 0, {}) as SummarizeOptions;
@@ -585,7 +585,7 @@ export async function execute(
 				item: index,
 			})),
 		};
-		return this.prepareOutputData([executionData]);
+		return [executionData];
 	} else {
 		if (!fieldsToSplitBy.length) {
 			const { pairedItems, ...json } = aggregationResult;
@@ -595,7 +595,7 @@ export async function execute(
 					item: index,
 				})),
 			};
-			return this.prepareOutputData([executionData]);
+			return [executionData];
 		}
 		const returnData = aggregationToArray(aggregationResult, fieldsToSplitBy);
 		const executionData = returnData.map((item) => {
@@ -607,6 +607,6 @@ export async function execute(
 				})),
 			};
 		});
-		return this.prepareOutputData(executionData);
+		return executionData;
 	}
 }
