+++ packages/nodes-base/nodes/ItemLists/V3/actions/router.ts
@@ -1,684 +1,35 @@
-import type { IDataObject, IExecuteFunctions, INodeExecutionData } from 'n8n-workflow';
-import { NodeOperationError, deepCopy } from 'n8n-workflow';
+import type { IExecuteFunctions, INodeExecutionData } from 'n8n-workflow';
+import { NodeOperationError } from 'n8n-workflow';
+import type { ItemListsType } from './node.type';
 
-import type { NodeVMOptions } from 'vm2';
-import { NodeVM } from 'vm2';
-
-import get from 'lodash/get';
-import isEmpty from 'lodash/isEmpty';
-import isEqual from 'lodash/isEqual';
-import lt from 'lodash/lt';
-import pick from 'lodash/pick';
-import set from 'lodash/set';
-import unset from 'lodash/unset';
-
-import * as summarize from './itemList/summarize.operation';
-import { compareItems, flattenKeys, shuffleArray } from '../helpers/utils';
+import * as itemList from './itemList';
 
 export async function router(this: IExecuteFunctions): Promise<INodeExecutionData[][]> {
+	let returnData: INodeExecutionData[] = [];
+
 	const items = this.getInputData();
-	const length = items.length;
-	const returnData: INodeExecutionData[] = [];
-	const resource = this.getNodeParameter('resource', 0);
+	const resource = this.getNodeParameter<ItemListsType>('resource', 0);
 	const operation = this.getNodeParameter('operation', 0);
-	const nodeVersion = this.getNode().typeVersion;
 
-	if (resource === 'itemList') {
-		if (operation === 'splitOutItems') {
-			for (let i = 0; i < length; i++) {
-				const fieldsToSplitOut = (this.getNodeParameter('fieldToSplitOut', i) as string)
-					.split(',')
-					.map((field) => field.trim());
-				const disableDotNotation = this.getNodeParameter(
-					'options.disableDotNotation',
-					0,
-					false,
-				) as boolean;
-
-				const destinationFields = (
-					this.getNodeParameter('options.destinationFieldName', i, '') as string
-				)
-					.split(',')
-					.filter((field) => field.trim() !== '')
-					.map((field) => field.trim());
-
-				if (destinationFields.length && destinationFields.length !== fieldsToSplitOut.length) {
-					throw new NodeOperationError(
-						this.getNode(),
-						'If multiple fields to split out are given, the same number of destination fields must be given',
-					);
-				}
-
-				const include = this.getNodeParameter('include', i) as
-					| 'selectedOtherFields'
-					| 'allOtherFields'
-					| 'noOtherFields';
-
-				const multiSplit = fieldsToSplitOut.length > 1;
-
-				const item = { ...items[i].json };
-				const splited: IDataObject[] = [];
-				for (const [entryIndex, fieldToSplitOut] of fieldsToSplitOut.entries()) {
-					const destinationFieldName = destinationFields[entryIndex] || '';
-
-					let arrayToSplit;
-					if (!disableDotNotation) {
-						arrayToSplit = get(item, fieldToSplitOut);
-					} else {
-						arrayToSplit = item[fieldToSplitOut];
-					}
-
-					if (arrayToSplit === undefined) {
-						if (nodeVersion < 2.1) {
-							if (fieldToSplitOut.includes('.') && disableDotNotation) {
-								throw new NodeOperationError(
-									this.getNode(),
-									`Couldn't find the field '${fieldToSplitOut}' in the input data`,
-									{
-										description:
-											"If you're trying to use a nested field, make sure you turn off 'disable dot notation' in the node options",
-									},
-								);
-							} else {
-								throw new NodeOperationError(
-									this.getNode(),
-									`Couldn't find the field '${fieldToSplitOut}' in the input data`,
-									{ itemIndex: i },
-								);
-							}
-						} else {
-							arrayToSplit = [];
-						}
-					}
-
-					if (typeof arrayToSplit !== 'object' || arrayToSplit === null) {
-						if (nodeVersion < 2.2) {
-							throw new NodeOperationError(
-								this.getNode(),
-								`The provided field '${fieldToSplitOut}' is not an array or object`,
-								{ itemIndex: i },
-							);
-						} else {
-							arrayToSplit = [arrayToSplit];
-						}
-					}
-
-					if (!Array.isArray(arrayToSplit)) {
-						arrayToSplit = Object.values(arrayToSplit);
-					}
-
-					for (const [elementIndex, element] of arrayToSplit.entries()) {
-						if (splited[elementIndex] === undefined) {
-							splited[elementIndex] = {};
-						}
-
-						const fieldName = destinationFieldName || fieldToSplitOut;
-
-						if (typeof element === 'object' && element !== null && include === 'noOtherFields') {
-							if (destinationFieldName === '' && !multiSplit) {
-								splited[elementIndex] = { ...splited[elementIndex], ...element };
-							} else {
-								splited[elementIndex][fieldName] = element;
-							}
-						} else {
-							splited[elementIndex][fieldName] = element;
-						}
-					}
-				}
-
-				for (const splitEntry of splited) {
-					let newItem: IDataObject = {};
-
-					if (include === 'noOtherFields') {
-						newItem = splitEntry;
-					}
-
-					if (include === 'allOtherFields') {
-						const itemCopy = deepCopy(item);
-						for (const fieldToSplitOut of fieldsToSplitOut) {
-							if (!disableDotNotation) {
-								unset(itemCopy, fieldToSplitOut);
-							} else {
-								delete itemCopy[fieldToSplitOut];
-							}
-						}
-						newItem = { ...itemCopy, ...splitEntry };
-					}
-
-					if (include === 'selectedOtherFields') {
-						const fieldsToInclude = (
-							this.getNodeParameter('fieldsToInclude.fields', i, []) as [{ fieldName: string }]
-						).map((field) => field.fieldName);
-
-						if (!fieldsToInclude.length) {
-							throw new NodeOperationError(this.getNode(), 'No fields specified', {
-								description: 'Please add a field to include',
-							});
-						}
-
-						for (const field of fieldsToInclude) {
-							if (!disableDotNotation) {
-								splitEntry[field] = get(item, field);
-							} else {
-								splitEntry[field] = item[field];
-							}
-						}
-
-						newItem = splitEntry;
-					}
-
-					returnData.push({
-						json: newItem,
-						pairedItem: {
-							item: i,
-						},
-					});
-				}
-			}
-
-			return this.prepareOutputData(returnData);
-		} else if (operation === 'concatenateItems') {
-			const aggregate = this.getNodeParameter('aggregate', 0, '') as string;
-
-			if (aggregate === 'aggregateIndividualFields') {
-				const disableDotNotation = this.getNodeParameter(
-					'options.disableDotNotation',
-					0,
-					false,
-				) as boolean;
-				const mergeLists = this.getNodeParameter('options.mergeLists', 0, false) as boolean;
-				const fieldsToAggregate = this.getNodeParameter(
-					'fieldsToAggregate.fieldToAggregate',
-					0,
-					[],
-				) as [{ fieldToAggregate: string; renameField: boolean; outputFieldName: string }];
-				const keepMissing = this.getNodeParameter('options.keepMissing', 0, false) as boolean;
-
-				if (!fieldsToAggregate.length) {
-					throw new NodeOperationError(this.getNode(), 'No fields specified', {
-						description: 'Please add a field to aggregate',
-					});
-				}
-
-				if (nodeVersion < 2.1) {
-					for (const { fieldToAggregate } of fieldsToAggregate) {
-						let found = false;
-						for (const item of items) {
-							if (fieldToAggregate === '') {
-								throw new NodeOperationError(this.getNode(), 'Field to aggregate is blank', {
-									description: 'Please add a field to aggregate',
-								});
-							}
-							if (!disableDotNotation) {
-								if (get(item.json, fieldToAggregate) !== undefined) {
-									found = true;
-								}
-							} else if (item.json.hasOwnProperty(fieldToAggregate)) {
-								found = true;
-							}
-						}
-						if (!found && disableDotNotation && fieldToAggregate.includes('.')) {
-							throw new NodeOperationError(
-								this.getNode(),
-								`Couldn't find the field '${fieldToAggregate}' in the input data`,
-								{
-									description:
-										"If you're trying to use a nested field, make sure you turn off 'disable dot notation' in the node options",
-								},
-							);
-						} else if (!found && !keepMissing) {
-							throw new NodeOperationError(
-								this.getNode(),
-								`Couldn't find the field '${fieldToAggregate}' in the input data`,
-							);
-						}
-					}
-				}
-
-				const newItem: INodeExecutionData = {
-					json: {},
-					pairedItem: Array.from({ length }, (_, i) => i).map((index) => {
-						return {
-							item: index,
-						};
-					}),
-				};
-
-				const values: { [key: string]: any } = {};
-				const outputFields: string[] = [];
-
-				for (const { fieldToAggregate, outputFieldName, renameField } of fieldsToAggregate) {
-					const field = renameField ? outputFieldName : fieldToAggregate;
-
-					if (outputFields.includes(field)) {
-						throw new NodeOperationError(
-							this.getNode(),
-							`The '${field}' output field is used more than once`,
-							{ description: 'Please make sure each output field name is unique' },
-						);
-					} else {
-						outputFields.push(field);
-					}
-
-					const getFieldToAggregate = () =>
-						!disableDotNotation && fieldToAggregate.includes('.')
-							? fieldToAggregate.split('.').pop()
-							: fieldToAggregate;
-
-					const _outputFieldName = outputFieldName
-						? outputFieldName
-						: (getFieldToAggregate() as string);
-
-					if (fieldToAggregate !== '') {
-						values[_outputFieldName] = [];
-						for (let i = 0; i < length; i++) {
-							if (!disableDotNotation) {
-								let value = get(items[i].json, fieldToAggregate);
-
-								if (!keepMissing) {
-									if (Array.isArray(value)) {
-										value = value.filter((entry) => entry !== null);
-									} else if (value === null || value === undefined) {
-										continue;
-									}
-								}
-
-								if (Array.isArray(value) && mergeLists) {
-									values[_outputFieldName].push(...value);
-								} else {
-									values[_outputFieldName].push(value);
-								}
-							} else {
-								let value = items[i].json[fieldToAggregate];
-
-								if (!keepMissing) {
-									if (Array.isArray(value)) {
-										value = value.filter((entry) => entry !== null);
-									} else if (value === null || value === undefined) {
-										continue;
-									}
-								}
-
-								if (Array.isArray(value) && mergeLists) {
-									values[_outputFieldName].push(...value);
-								} else {
-									values[_outputFieldName].push(value);
-								}
-							}
-						}
-					}
-				}
-
-				for (const key of Object.keys(values)) {
-					if (!disableDotNotation) {
-						set(newItem.json, key, values[key]);
-					} else {
-						newItem.json[key] = values[key];
-					}
-				}
-
-				returnData.push(newItem);
-
-				return this.prepareOutputData(returnData);
-			} else {
-				let newItems: IDataObject[] = items.map((item) => item.json);
-				const destinationFieldName = this.getNodeParameter('destinationFieldName', 0) as string;
-				const fieldsToExclude = (
-					this.getNodeParameter('fieldsToExclude.fields', 0, []) as IDataObject[]
-				).map((entry) => entry.fieldName);
-				const fieldsToInclude = (
-					this.getNodeParameter('fieldsToInclude.fields', 0, []) as IDataObject[]
-				).map((entry) => entry.fieldName);
-
-				if (fieldsToExclude.length || fieldsToInclude.length) {
-					newItems = newItems.reduce((acc, item) => {
-						const newItem: IDataObject = {};
-						let outputFields = Object.keys(item);
-
-						if (fieldsToExclude.length) {
-							outputFields = outputFields.filter((key) => !fieldsToExclude.includes(key));
-						}
-						if (fieldsToInclude.length) {
-							outputFields = outputFields.filter((key) =>
-								fieldsToInclude.length ? fieldsToInclude.includes(key) : true,
-							);
-						}
-
-						outputFields.forEach((key) => {
-							newItem[key] = item[key];
-						});
-
-						if (isEmpty(newItem)) {
-							return acc;
-						}
-						return acc.concat([newItem]);
-					}, [] as IDataObject[]);
-				}
-
-				const output: INodeExecutionData = { json: { [destinationFieldName]: newItems } };
-
-				return this.prepareOutputData([output]);
-			}
-		} else if (operation === 'removeDuplicates') {
-			const compare = this.getNodeParameter('compare', 0) as string;
-			const disableDotNotation = this.getNodeParameter(
-				'options.disableDotNotation',
-				0,
-				false,
-			) as boolean;
-			const removeOtherFields = this.getNodeParameter(
-				'options.removeOtherFields',
-				0,
-				false,
-			) as boolean;
-
-			let keys = disableDotNotation
-				? Object.keys(items[0].json)
-				: Object.keys(flattenKeys(items[0].json));
-
-			for (const item of items) {
-				for (const key of disableDotNotation
-					? Object.keys(item.json)
-					: Object.keys(flattenKeys(item.json))) {
-					if (!keys.includes(key)) {
-						keys.push(key);
-					}
-				}
-			}
-
-			if (compare === 'allFieldsExcept') {
-				const fieldsToExclude = (
-					this.getNodeParameter('fieldsToExclude.fields', 0, []) as [{ fieldName: string }]
-				).map((field) => field.fieldName);
-				if (!fieldsToExclude.length) {
-					throw new NodeOperationError(
-						this.getNode(),
-						'No fields specified. Please add a field to exclude from comparison',
-					);
-				}
-				if (!disableDotNotation) {
-					keys = Object.keys(flattenKeys(items[0].json));
-				}
-				keys = keys.filter((key) => !fieldsToExclude.includes(key));
-			}
-			if (compare === 'selectedFields') {
-				const fieldsToCompare = (
-					this.getNodeParameter('fieldsToCompare.fields', 0, []) as [{ fieldName: string }]
-				).map((field) => field.fieldName);
-				if (!fieldsToCompare.length) {
-					throw new NodeOperationError(
-						this.getNode(),
-						'No fields specified. Please add a field to compare on',
-					);
-				}
-				if (!disableDotNotation) {
-					keys = Object.keys(flattenKeys(items[0].json));
-				}
-				keys = fieldsToCompare.map((key) => key.trim());
-			}
-
-			// This solution is O(nlogn)
-			// add original index to the items
-			const newItems = items.map(
-				(item, index) =>
-					({
-						json: { ...item.json, __INDEX: index },
-						pairedItem: { item: index },
-					} as INodeExecutionData),
-			);
-			//sort items using the compare keys
-			newItems.sort((a, b) => {
-				let result = 0;
-
-				for (const key of keys) {
-					let equal;
-					if (!disableDotNotation) {
-						equal = isEqual(get(a.json, key), get(b.json, key));
-					} else {
-						equal = isEqual(a.json[key], b.json[key]);
-					}
-					if (!equal) {
-						let lessThan;
-						if (!disableDotNotation) {
-							lessThan = lt(get(a.json, key), get(b.json, key));
-						} else {
-							lessThan = lt(a.json[key], b.json[key]);
-						}
-						result = lessThan ? -1 : 1;
-						break;
-					}
-				}
-				return result;
-			});
-
-			for (const key of keys) {
-				let type: any = undefined;
-				for (const item of newItems) {
-					if (key === '') {
-						throw new NodeOperationError(this.getNode(), 'Name of field to compare is blank');
-					}
-					const value = !disableDotNotation ? get(item.json, key) : item.json[key];
-					if (value === undefined && disableDotNotation && key.includes('.')) {
-						throw new NodeOperationError(
-							this.getNode(),
-							`'${key}' field is missing from some input items`,
-							{
-								description:
-									"If you're trying to use a nested field, make sure you turn off 'disable dot notation' in the node options",
-							},
-						);
-					} else if (value === undefined) {
-						throw new NodeOperationError(
-							this.getNode(),
-							`'${key}' field is missing from some input items`,
-						);
-					}
-					if (type !== undefined && value !== undefined && type !== typeof value) {
-						throw new NodeOperationError(this.getNode(), `'${key}' isn't always the same type`, {
-							description: 'The type of this field varies between items',
-						});
-					} else {
-						type = typeof value;
-					}
-				}
-			}
-
-			// collect the original indexes of items to be removed
-			const removedIndexes: number[] = [];
-			let temp = newItems[0];
-			for (let index = 1; index < newItems.length; index++) {
-				if (compareItems(newItems[index], temp, keys, disableDotNotation, this.getNode())) {
-					removedIndexes.push(newItems[index].json.__INDEX as unknown as number);
-				} else {
-					temp = newItems[index];
-				}
-			}
-
-			let data = items.filter((_, index) => !removedIndexes.includes(index));
-
-			if (removeOtherFields) {
-				data = data.map((item, index) => ({
-					json: pick(item.json, ...keys),
-					pairedItem: { item: index },
-				}));
-			}
-
-			// return the filtered items
-			return this.prepareOutputData(data);
-		} else if (operation === 'sort') {
-			let newItems = [...items];
-			const type = this.getNodeParameter('type', 0) as string;
-			const disableDotNotation = this.getNodeParameter(
-				'options.disableDotNotation',
-				0,
-				false,
-			) as boolean;
-
-			if (type === 'random') {
-				shuffleArray(newItems);
-				return this.prepareOutputData(newItems);
-			}
-
-			if (type === 'simple') {
-				const sortFieldsUi = this.getNodeParameter('sortFieldsUi', 0) as IDataObject;
-				const sortFields = sortFieldsUi.sortField as Array<{
-					fieldName: string;
-					order: 'ascending' | 'descending';
-				}>;
-
-				if (!sortFields?.length) {
-					throw new NodeOperationError(
-						this.getNode(),
-						'No sorting specified. Please add a field to sort by',
-					);
-				}
-
-				for (const { fieldName } of sortFields) {
-					let found = false;
-					for (const item of items) {
-						if (!disableDotNotation) {
-							if (get(item.json, fieldName) !== undefined) {
-								found = true;
-							}
-						} else if (item.json.hasOwnProperty(fieldName)) {
-							found = true;
-						}
-					}
-					if (!found && disableDotNotation && fieldName.includes('.')) {
-						throw new NodeOperationError(
-							this.getNode(),
-							`Couldn't find the field '${fieldName}' in the input data`,
-							{
-								description:
-									"If you're trying to use a nested field, make sure you turn off 'disable dot notation' in the node options",
-							},
-						);
-					} else if (!found) {
-						throw new NodeOperationError(
-							this.getNode(),
-							`Couldn't find the field '${fieldName}' in the input data`,
-						);
-					}
-				}
-
-				const sortFieldsWithDirection = sortFields.map((field) => ({
-					name: field.fieldName,
-					dir: field.order === 'ascending' ? 1 : -1,
-				}));
-
-				newItems.sort((a, b) => {
-					let result = 0;
-					for (const field of sortFieldsWithDirection) {
-						let equal;
-						if (!disableDotNotation) {
-							const _a =
-								typeof get(a.json, field.name) === 'string'
-									? (get(a.json, field.name) as string).toLowerCase()
-									: get(a.json, field.name);
-							const _b =
-								typeof get(b.json, field.name) === 'string'
-									? (get(b.json, field.name) as string).toLowerCase()
-									: get(b.json, field.name);
-							equal = isEqual(_a, _b);
-						} else {
-							const _a =
-								typeof a.json[field.name] === 'string'
-									? (a.json[field.name] as string).toLowerCase()
-									: a.json[field.name];
-							const _b =
-								typeof b.json[field.name] === 'string'
-									? (b.json[field.name] as string).toLowerCase()
-									: b.json[field.name];
-							equal = isEqual(_a, _b);
-						}
-
-						if (!equal) {
-							let lessThan;
-							if (!disableDotNotation) {
-								const _a =
-									typeof get(a.json, field.name) === 'string'
-										? (get(a.json, field.name) as string).toLowerCase()
-										: get(a.json, field.name);
-								const _b =
-									typeof get(b.json, field.name) === 'string'
-										? (get(b.json, field.name) as string).toLowerCase()
-										: get(b.json, field.name);
-								lessThan = lt(_a, _b);
-							} else {
-								const _a =
-									typeof a.json[field.name] === 'string'
-										? (a.json[field.name] as string).toLowerCase()
-										: a.json[field.name];
-								const _b =
-									typeof b.json[field.name] === 'string'
-										? (b.json[field.name] as string).toLowerCase()
-										: b.json[field.name];
-								lessThan = lt(_a, _b);
-							}
-							if (lessThan) {
-								result = -1 * field.dir;
-							} else {
-								result = 1 * field.dir;
-							}
-							break;
-						}
-					}
-					return result;
-				});
-			} else {
-				const code = this.getNodeParameter('code', 0) as string;
-				const regexCheck = /\breturn\b/g.exec(code);
-
-				if (regexCheck?.length) {
-					const sandbox = {
-						newItems,
-					};
-					const mode = this.getMode();
-					const options = {
-						console: mode === 'manual' ? 'redirect' : 'inherit',
-						sandbox,
-					};
-					const vm = new NodeVM(options as unknown as NodeVMOptions);
-
-					newItems = await vm.run(
-						`
-					module.exports = async function() {
-						newItems.sort( (a,b) => {
-							${code}
-						})
-						return newItems;
-					}()`,
-						__dirname,
-					);
-				} else {
-					throw new NodeOperationError(
-						this.getNode(),
-						"Sort code doesn't return. Please add a 'return' statement to your code",
-					);
-				}
-			}
-			return this.prepareOutputData(newItems);
-		} else if (operation === 'limit') {
-			let newItems = items;
-			const maxItems = this.getNodeParameter('maxItems', 0) as number;
-			const keep = this.getNodeParameter('keep', 0) as string;
-
-			if (maxItems > items.length) {
-				return this.prepareOutputData(newItems);
-			}
-
-			if (keep === 'firstItems') {
-				newItems = items.slice(0, maxItems);
-			} else {
-				newItems = items.slice(items.length - maxItems, items.length);
-			}
-			return this.prepareOutputData(newItems);
-		} else if (operation === 'summarize') {
-			return summarize.execute.call(this, items);
-		} else {
-			throw new NodeOperationError(this.getNode(), `Operation '${operation}' is not recognized`);
+	const itemListsNodeData = {
+		resource,
+		operation,
+	} as ItemListsType;
+
+	try {
+		switch (itemListsNodeData.resource) {
+			case 'itemList':
+				returnData = await itemList[itemListsNodeData.operation].execute.call(this, items);
+				break;
+			default:
+				throw new NodeOperationError(
+					this.getNode(),
+					`The operation "${operation}" is not supported!`,
+				);
 		}
-	} else {
-		throw new NodeOperationError(this.getNode(), `Resource '${resource}' is not recognized`);
+	} catch (error) {
+		throw error;
 	}
+
+	return this.prepareOutputData(returnData);
 }
